correctly add HAL lib names to PRODUCT_PACKAGES
libhardware uses ro.product.board and ro.board.platform to look for
HAL names. ro.board.platform comes from TARGET_BOARD_PLATFORM,
enforce that it is set this early in the build process.

PRODUCT_DEVICE is not appropriate to use as it may not be set at
the time the product-level Makefile is read if it is set in a different
file which calls inherit-product, and libhardware doesn't actually
use it anyway.

Change-Id: Idd3c28e07ee45ef3cfc367dad35a97529fb68274
Signed-off-by: Andrew Boie <andrew.p.boie@intel.com>
Reviewed-on: https://android.intel.com/203494
Reviewed-by: Gumbel, Matthew K <matthew.k.gumbel@intel.com>
Reviewed-by: Ross, Andrew J <andrew.j.ross@intel.com>
Reviewed-by: Uyyala, Sridhar <sridhar.uyyala@intel.com>