houdini: check for houdini in both hardware/ and vendor/
Houdini contains an XML module called check.xml.  After conversion to a
"binary blob", this file will be installed in the user's source tree via
PRODUCT_COPY_FILES. Android build system requires that all XML files
installed via PRODUCT_COPY_FILES pass xmllint. Problem is, check.xml is not
valid XML. This issue has been raised to the Houdini team, but this patch
is not a REVERTME because vendor/ is not the right place for prebuilt blobs
and really this project should live in hardware/.

The reason we check both hardware/ and vendor/ is because the manifest
patch will not land on master until after this patch has landed on build
topic branch. We need to check for both here so that build branch will not
be broken. Later, after this change lands on master, we can remove the
vendor/ part.

Depends-Change-Id: Ia809d358502d7dd24e2cfcea32af52a4c5d6c0c6
Change-Id: I2fb724948f83f16c4352bfb71188b10b5dac509d
Signed-off-by: Matt Gumbel <matthew.k.gumbel@intel.com>
Reviewed-on: https://android.intel.com/209689