telemetry: Change service classes according to lifecycle.
Change tm_dmesg_probe and tm_pstore_probe to class core, and telemetryd
to class late_start.

In case of tm_dmesg_probe that's because there's no need to stop and
restart the process when /data is mounted, it doesn't use /data.

In case of tm_pstore_probe and telemetryd since they're "disabled" it
should make no practical difference but these classes match when the
services start and stop.

Change-Id: I1ade5af4dfbdb9cad003fc86f37a1cc359806e69
Signed-off-by: Andrew Zaborowski <andrew.zaborowski@intel.com>