STPK-407: Call the telemetry storage setup script in init.rc
The tm_storage_setup script is called in post-fs, post-fs-data and on
telemetry being enabled or disabled through the persist.intel.tm.state
property.  Because we only use the /cache datastore before encryption
is set up by vold, the setup of the /data datastore depends on the state
of more than one property so the logic to disable and enable that
datastore became too complex to manage in just the init language's
triggers.  Instead we call tm_storage_setup when either of these
properties change and let the script figure out which datastores should
be enabled or disabled based on the properties' values..

In addition to the way the datastores are being set up there is a
functional change in that the unsubmitted record queues are now removed
when telemetry gets disabled by user, together with the directory image
(includes machine_id).

Issue: STPK-407
Change-Id: I6653805830dd341da404b33f46564b3cc28b84cb
Signed-off-by: Andrew Zaborowski <andrew.zaborowski@intel.com>