pstore: create mixins group to handle pstore backends
pstore can be used by only one backend at a time. By default this is
handled by the kernel, the 1st backend to register will be the one to
be used. Currently in our kernel this is the efi backend.

As we need to use pstore ram backend for stabilization, provides
runtime console and ftrace support, we need to have a way to select
which backend is used.

This can be solved by passing the requested backend to kernel through
the cmdline.

This pstore mixins group introduce this feature. See README for
details on each option.

Change-Id: I7126397192cf0414b352eaeaf47da9e2efbd6102
Tracked-On: https://jira01.devtools.intel.com/browse/GMINL-9360
Signed-off-by: Jeremy Rocher <jeremy.rocher@intel.com>
Signed-off-by: Nicolae Natea <nicolaex.natea@intel.com>
Reviewed-on: https://android.intel.com/372515