telemetry: include log probe in userdebug only
We remove the special "true_with_logprobe" mixin and add a check in the
"true" section to only include the log probe and its configuration file
in debug builds. We keep the logprobe related control in the init.rc file
no matter if the binary is present or not.
A new logevent configuration file is installed during this time - only if
in userdebug image.

Change-Id: I1edeaabe80e432cd0b04c5d9ddcfc652e401b180
Signed-off-by: Reinette Chatre <reinette.chatre@intel.com>
Reviewed-on: https://android.intel.com:443/299668