gps: Remove gpsd from the service main class
The gpsd service is now started by hald upon device detection and is
set as 'disabled' in init.rc to not start automatically at boot time.

This patch removes it from the service main class because main class
services are reset after a factory boot and gpsd is not restarted thus
because specified as disabled.

Issue: IRDA-4725
Change-Id: I5d9a32afef4728aa5f7e66091cca22afec1be35d
Signed-off-by: Thierry Escande <thierry.escande@linux.intel.com>
Reviewed-on: https://android.intel.com/311373
Reviewed-by: Poulain, Loic <loic.poulain@intel.com>
Tested-by: Poulain, Loic <loic.poulain@intel.com>
Reviewed-by: Escande, Thierry <thierry.escande@intel.com>
Tested-by: Escande, Thierry <thierry.escande@intel.com>