telemetry: start log probe by default
The telemetry log probe (tm_logprobe) can now also support privacy compliant
events. Which events are considered as such are specified in
/etc/tm/logevent/event-log/tags that is not writable. We now start the log
probe by default and reduce the priviledges needed for it to operate. The
log probe is started as user telemetry, for which we now need to add
read_logd() permissions to its SELinux policy. We also need to add the
group ids of cache and log for it to be able to operate.

As a start we only support the battery_level compliant event.

Change-Id: If468c3322881c4cc21a001bc71f544e78049b534
Signed-off-by: Reinette Chatre <reinette.chatre@intel.com>