[TELEPHONY/MIXIN]: use ssic hsic port to disable ipc links
Because the IPC links connected to the modem must be
disable before MMGR service starts, it has to be done
in the init.rc and ueventd.rc files. But as this is
dedicated to the telephony, it has to be implemented
in the mixin telephony group by using the variables
defined in the mixin.spec file of the platform.

Change-Id: I82f6f4a2afa076b9b722a6c0fbd70d8ef83dfe47
Depends-Change-Id: Ic19581be29e735586ac440bc16994dc5c9c2c94a
Signed-off-by: Gwenn Bourree <gwenn.bourree@intel.com>
Tracked-On: https://jira01.devtools.intel.com/browse/GMINL-4805
Reviewed-on: https://android.intel.com/313015
Reviewed-by: Svenberg, Stefan <stefan.svenberg@intel.com>
Tested-by: Svenberg, Stefan <stefan.svenberg@intel.com>