device-type: Add autodetect type
In order to reduce the number of permissions installed by default,
we can not use one of the hardcoded AOSP (tablet, clamshell or
handheld) XML file. For example the AOSP tablet one installs the
compass feature by default although some of the IRDA SKUs do not
have it.
So we're creating a new 'autodetect' device type, that installs a
minimal (although it can still be reduced) core_hardware.xml file.

Change-Id: I8fcd4fee47cca014ec90a64d4269a812fc106b2b
Signed-off-by: Samuel Ortiz <sameo@linux.intel.com>
Reviewed-on: https://android.intel.com/316467
Reviewed-by: cactus <cactus@intel.com>
Reviewed-by: Escande, Thierry <thierry.escande@intel.com>
Reviewed-by: Nashif, Anas <anas.nashif@intel.com>
Reviewed-by: Mathew, Alwin <alwin.mathew@intel.com>
Tested-by: Mathew, Alwin <alwin.mathew@intel.com>