Change parameter substitution from string.Formatter to pystache
Change the parameter substution engine to pystache which
implements the "Mustache" templating language. This allows a
richer templating syntax. This patch changes just the template
formatting. The format for accepting parameters from the
mixins.spec is not updated to allow richer parameters with one
exception. A parameter value of 'False' will be converted to the
boolean value False and so will work as a false value in the
template.

For more information on Mustache templates, see the reference:

http://mustache.github.io/mustache.5.html

Signed-off-by: Scott D Phillips <scott.d.phillips@intel.com>
Change-Id: Iac5aa82ab9d3e308ba61a5e09ed4cbbb8232d0db
Signed-off-by: Radivoje Jovanovic <radivoje.jovanovic@intel.com>