autodetect: Use GMin's Audio HAL again.
This commit bring back the GMin Audio HAL onto coho and cohol builds along with
the relevant configuration files.

For now, we keep the irda hal in the build but GMin HAL support for all boards
will be gradualy added and the irda hal will be eventually removed.

Issue: GMINL-6770

Change-Id: Iebe680d909fa2dc128e6c224eec3dd0a1207de29
Signed-off-by: David Wagner <david.wagner@intel.com>
Reviewed-on: https://android.intel.com/341870
Reviewed-by: Bohlin, Johan <johan.bohlin@intel.com>
Tested-by: Bohlin, Johan <johan.bohlin@intel.com>