audio: Add HAL build configuration flag
A build flag is needed to decide which audio hal that
shall be built. This is needed because the HAL has
to have certain names (decided by android) to get loaded,
unless autodetect is used. The effect of this is that all
hals (not using autodetect) has to have module set to
LOCAL_MODULE := audio.primary.$(TARGET_BOARD_PLATFORM).
To separate which hal that is actually built with a certain
device we have introduced a build flag.

Issue: GMINL-2900

Change-Id: I5e846c459633094b23fdd2dedc4f8899d6b7c9a7
Signed-off-by: johan.bohlin <johan.bohlin@intel.com>
Reviewed-on: https://android.intel.com/355305
Reviewed-by: Lilja, Ola <ola.lilja@intel.com>