shrink bootloader and cache partitions and filesystems
- The original direction on what size to use for /cache was from
a Google engineer who suggested over email 256M. However, once
the CDD came out the recommended size is now 100M.

- The bootloader partition (plus the backup work area) was sized
at 100M, but this is too generous based on actual usage. On ecs_e7
only about 25M are being used. Reduce the size to 60M which still
has plenty of expansion room.

The effect of this patch on devices will be as follows:

- For new devices, there will be 236M of additional space for user
data.
- For existing devices, the filesystems will be smaller, but the
unused space at the end of the partitions will just be dead bits.

Change-Id: Id4b6d1c93046894df4bdb00550db4a65ab3a6e44
Signed-off-by: Andrew Boie <andrew.p.boie@intel.com>
Reviewed-on: https://android.intel.com/337151
Reviewed-by: Afantenos, Marie Cecile <marie.cecile.afantenos@intel.com>
Reviewed-by: Compostella, Jeremy <jeremy.compostella@intel.com>
Reviewed-by: Lynch, Rusty <rusty.lynch@intel.com>
Reviewed-by: Mahalingam, Ganesh <ganesh.mahalingam@intel.com>