Revert "Revert "increase size of metadata partition""
Confirmed not to be an error. Google wants to store additional
data here beyond the crypto footer.

"We need to be able to show some info when the device is locked
(owner (name, image,...), language, timezone, ...). We want
extra backup for crypto and verity info (e.g. dm-verity hashes
for /system are: 5MB)"

This reverts commit eb92042c63dfd4558f9cec6d803d37a7c5c2bcb4.

Change-Id: I67453e2d7af6156369918548cb1a312ebdbea81a
Signed-off-by: Andrew Boie <andrew.p.boie@intel.com>
Reviewed-on: https://android.intel.com/341646
Reviewed-by: Compostella, Jeremy <jeremy.compostella@intel.com>
Reviewed-by: Geltz, Brad <brad.geltz@intel.com>