telephony: add MMGR client library
Modem Management API has been updated (C and java).

All modem clients are linked to proxy library which loads the
right implementation according to scalability parameters.

For C world, the proxy library is called libmdmcli.so.
The name of the library for MMGR is libmdmcli_mmgr.so.
Scalability parameter is retrieved with TCS.

For java world the proxy is called
com.intel.internal.telephony.ModemClient. The name of the library
for MMGR is com.intel.internal.telephony.MmgrClient.
Scalability parameter is retrieved via two properties, set via
the catalog.

With this new mechanism, libraries need to be declared
in the mixin, otherwise they are not embedded in the platform.

This patch adds those two dependencies and update the catalog
with the new properties

Change-Id: I06ad2d25935a5a2c3e431ca47556eca90ff57d1d
Tracked-On: https://jira01.devtools.intel.com/browse/IMINAN-19136
Signed-off-by: Cesar DE OLIVEIRA <cesar.de.oliveira@intel.com>
Reviewed-on: https://android.intel.com:443/350001
Reviewed-on: https://android.intel.com/363063
Reviewed-by: Paut, Frederic <frederic.paut@intel.com>
Reviewed-by: Svenberg, Stefan <stefan.svenberg@intel.com>
Reviewed-by: Nunes, Philippe <philippe.nunes@intel.com>
Tested-by: Nunes, Philippe <philippe.nunes@intel.com>