telephony: change access rights for M.2 USB-HS interfaces
In order to correctly flash the M.2 modem the usb2_hardware_lpm
must be disabled from MMGR before flashing.
In order to enable LPM for the M.2 modem the power/control must
be set to auto from MMGR after flashing the modem.
This patch changes the groups and the rights for the two sysfs
so they can be accessed from MMGR.

Change-Id: I07ef500d4f231d7756ffd957c9bd7664259a36bf
Tracked-On: https://jira01.devtools.intel.com/browse/IMINAN-33423
Signed-off-by: George Sandu <georgex.sandu@intel.com>
Reviewed-on: https://android.intel.com:443/357130
Reviewed-on: https://android.intel.com/362894
Reviewed-by: Svenberg, Stefan <stefan.svenberg@intel.com>
Tested-by: Svenberg, Stefan <stefan.svenberg@intel.com>