Aligned mixin with BoardConfig
Previously there were entried in the top of the BoardConfig.mk
for the s3gr10m6s target which would override the values specified
in the mixin.  This change aligns the mixin values with the working
values in the BoardConfig.mk.

Once this is complete, additional changes will be necessary to
arrive at the correct partition sizing.

Change-Id: Ia6ecb952848f32bbcbb66c0062e29f6780e461a6
Signed-off-by: Brad Geltz <brad.geltz@intel.com>
Reviewed-on: https://android.intel.com/373465
Reviewed-by: Bergeron, Michael <michael.bergeron@intel.com>
Reviewed-by: Dattatraya Kulkarni, Anisha <anisha.dattatraya.kulkarni@intel.com>
Reviewed-by: Wu, Ji J <ji.j.wu@intel.com>