boot-arch: sofia: Do not exclude modem.fls from flashfiles
modem.fls is a separate prebuilt on LTE while the actual
modem firmware is embedded into the 3GR system.fls.

In the LTE case, excluding modem.fls from the flashfiles
generates a non bootable image. In the 3GR case, there is
no modem.fls.
In both cases there is no need to exclude it from the
flashfiles output.

Change-Id: I2d54f75fe80e2c72e8f6177c07c2b7d1c4ad02ca
Signed-off-by: Samuel Ortiz <sameo@linux.intel.com>
Reviewed-on: https://android.intel.com:443/375362