Update json format to v3.0 and add a mixin to check battery level
Json flashfiles are updated to version 3.0 to support new
functionalities such as parameters and preconditions.

Add a battery parameter to retrieve the battery level through a
fastboot command. Add also a battery precondition before each fastboot
command during flash to check the battery level. If it is not high
enough, the command will fail.

The battery parameter and precondition are enabled through the
fastboot_min_battery_level mixin which is set to false by default.

Change-Id: Id9a8c8c567663adfa4db98fb0de22b076b3ac6f2
Tracked-On: https://jira01.devtools.intel.com/browse/OAM-1618
Signed-off-by: Gaelle Nassiet <gaellex.nassiet@intel.com>
Reviewed-on: https://android.intel.com:443/406800