ota:fix recovery partition number fix recovery update issue
The issue here was that the fstab for the recovery partition was never
correct in the 1A tree.  Because no mechanism ever exercised the fstab
entry for the recovery partition, this was not found.  Xihua found this
when he was debugging the issue of the recovery partition not updating
via OTA.  The AOSP script  that performs the update, install-recovery.sh
(part of ./build/tools/releasetools/common.py) was trying to mount
ImcPartID072 which is not correct.  The change to ImcPartID121 also
aligns the 1A fstab declaration for the recovery partition with the
OC6 tree's version.

Change-Id: I8d4ba9a58ac37bf54c9be3e179396f4ff4758ae2
Tracked-On: https://jira01.devtools.intel.com/browse/GMINL-11523
Signed-off-by: xihua.chen <xihua.chen@intel.com>
Reviewed-on: https://android.intel.com/388668
Reviewed-by: Geltz, Brad <brad.geltz@intel.com>
Tested-by: Geltz, Brad <brad.geltz@intel.com>