[Telephony][sf3g-sflte] Add imc OemTelephony application
Add the imc OemTelephony application package to the Sofia LTE
and 3GR products.
There are 2 repositories for OemTelephony:
vendor/intel/apps/telephony/OemTelephony: The HOME for this
is 1A. Used for non-Sofia targets
vendor/intel/apps/telephony/OemTelephony-imc: The HOME for
this is OC6. Used for Sofia targets.

A recent change has modified the way to select and add imc
Oemtelephony application to the sf3g-sflte preventing it to
be included in the system images.
This patch fixes by updating the sf3g-sflte telephony mixins.

BOARD_USE_IMC_OEM_TELEPHONY=true has to be defined and the "imc"
suffix has to be removed from the related product packages name.

Tracked-On: https://jira01.devtools.intel.com/browse/GMINL-8151
Signed-off-by: cguiraud <christophe.guiraud@intel.com>
Change-Id: Ibcef265bb5aed3ac5a560ed04e7b899772754ce6
Depends-Change-Id: Iedbe71525b5ba939ed43aa2725d208409659fb26
Depends-Change-Id: I63c78de468ffcc3ec386ae2dcf662486975ddb58
Reviewed-on: https://android.intel.com:443/392569