flashfiles: allow format-data in user build
format-data (erase + format) command is disabled for user builds
because the lock command will erase the data partition and the
filesystem is lost.

However, if the device is in provisioning mode, the
unlock/verified/lock commands do not erase the data partition and
leave the data partition as-is.  When Android starts it tries to
uncrypt this non-empty data partition, fails and then prompts the user
for the key password.

This patch allows the format-data command in user build:
- in factory, the data partition will be erased and formatted which
  will also improve the first time boot.  The user build will boot
  without prompting the user for key password it does not have anyway.
- for engineering, it will slow down the flash time by a couple of
  seconds.

Change-Id: Ifc7f11c59e49918e835b9f8b0c454203fb2c110a
Tracked-On: https://jira01.devtools.intel.com/browse/GMINL-14977
Signed-off-by: Jeremy Compostella <jeremy.compostella@intel.com>
Reviewed-on: https://android.intel.com:443/395472