boot-arch/efi: add mixin parameter data_encryption
Default value is true.
If data_encryption is set to false, the encryption will be disabled on
/data from fstab and the partition will be marked as formattable.
If a wiped partition is neither encryptable nor formattable, it will
stay unformatted, hence unusable.
/data partition is set as formattable when encryption is disabled, so
if it is wiped during flash process or during a device state transition,
fs_mgr will format it at next boot.

Change-Id: Ie5deb13c3229b7269f0757b9269840544fe0f65c
Signed-off-by: Leo Sartre <leox.sartre@intel.com>
Tracked-On: https://jira01.devtools.intel.com/browse/OAM-85
Reviewed-on: https://android.intel.com:443/397888