boot-arch: configure watchdog interval/margin via mixin
A new 'watchdog_parameters' option is introduced in boot-arch mixin
group so that the watchdog 'interval' and 'margin' values can be
configured via mixin.
By default it is set to false, and in that case the default watchdog
interval (10s) and margin (10s) are used.
User can now change watchdog interval and magin values for a given
target by changing this option in the appropriate mixin.spec

Change-Id: I83597c88f7af6b946bbe0902585daf24fa90095d
Signed-off-by: Yann Fouassier <yann.fouassier@intel.com>
Tracked-On: https://jira01.devtools.intel.com/browse/OAM-501
Reviewed-on: https://android.intel.com:443/398059
Reviewed-on: https://android.intel.com:443/400962