[CWS_WIDI_UX] BCM4356 Do not scan in background when P2P connected/ing.
When no WiFi network is remembered, wpa_supplicant do not
automatically schedule periodic scan.
In that case, Android framework take the action to trig a scan
with the same interval of time than "suplicant scan interval".
When P2P is connected, this Android framework scan is not triggered
and supplicant scan interval is changed.
Set it to 3600s in order to minimize scan impact on P2P
connection, mainly for Miracast purpose.

Change-Id: I7bfeddf6d042c765015e4ca681fbaca79fb38039
Tracked-On: https://jira01.devtools.intel.com/browse/OAM-3717
Reviewed-on: https://android.intel.com:443/404191
Signed-off-by: Jean Trivelly <jean.trivelly@intel.com>
Signed-off-by: Aymen Ben Alaya <aymenx.ben.alaya@intel.com>
Reviewed-on: https://android.intel.com:443/415216