[CWS_WIDI_UX] iwlwifi. Adapter reconnection upon AP just connected on another frequency
Since some WiFi chip has some limitation to offer high video quality while used
in muli-roles and multi-channels, the best solution is to use the same channel as
the access point when running in multi-roles.

This solution is already in place when DUT is first connected to an access point
and then connected to a video adapter. However, in some case we can imagagine that
user first connects to video adapter and then connects to an access point. In
such situation, there is a high probability that adapter and access point
doesn't run on the same channel.

This patch detects when DUT is reaching an access point with a different
frequency than video adapter, a dialog box informs the user that the video link
will be closed and restored for better quality. The user has no other choice
than accept and the patch manages the disconnection, retrieves the adapter we
left and finally performs the reconnection.

Change-Id: Id98e65bb51308d51674ca4ed0267b19d6cd0ac50
Tracked-On: https://jira01.devtools.intel.com/browse/OAM-6989
Reviewed-on: https://android.intel.com:443/404193
Signed-off-by: Jean Trivelly <jean.trivelly@intel.com>
Signed-off-by: Aymen Ben Alaya <aymenx.ben.alaya@intel.com>
Reviewed-on: https://android.intel.com:443/430203