Create storage configuration for sdcard in mmc0 slot and USB storages.
For some platforms (for example BXT) it may happen that due to specific
BDF order SD HC may get enumerated as mmc0. As result of that, SD card
in this slot get enumerated as mmcblk0 or mmcblk1 depending on it being
present in slot during system boot or not. SD card is also marked as
adoptable in this configuration.

Also, if UFS storage is available it volumes get enumerated as sd*,
which may conflict with unproper USB configuration, so we explicitly
need to set in fstab that the enumeration path will be for USB storage.
This also has no limit of number of USB storages that can be handled.

Change-Id: I561c4dd480676ee12ad666aa1662e3d2ce014a46
Tracked-On: https://jira01.devtools.intel.com/browse/OAM-379
Tracked-On: https://jira01.devtools.intel.com/browse/OAM-610
Signed-off-by: Mateusz Nowak <mateusz.nowak@intel.com>
Reviewed-on: https://android.intel.com:443/404260