flashfiles: add ini-sofia option
The ini-sofia option symlinks to the ini option for the bulk of
its makefile logic, but leaves out the fastboot-flashing specific
flashfiles.ini template.

Also a previously missing dependency on
BOARD_FLASHFILES_FIRMWARE_$(variant) is added to provdata.zip in
the ini AndroidBoard.mk so that the build won't try to generate
provdata.zip before all its dependencies are generated.

Rebased patch set 6.

Change-Id: I0ef9b6672abd138f1bcda7fb456dfe14eb3e6e7d
Tracked-On: https://jira01.devtools.intel.com/browse/OAM-4058
Signed-off-by: Scott D Phillips <scott.d.phillips@intel.com>
Reviewed-on: https://android.intel.com:443/410052