[CWS_WIDI_UX] autodetect-sofia. Do not scan in background when P2P connected/ing.
When no WiFi network is remembered, wpa_supplicant do not
automatically schedule periodic scan.
In that case, Android framework take the action to trig a scan
with the same interval of time than "suplicant scan interval".
When P2P is connected, this Android framework scan is not triggered
and supplicant scan interval is changed.
Set it to 3600s in order to minimize scan impact on P2P
connection, mainly for Miracast purpose.

Change-Id: I22ccb2c13d2007c675d798f48b2f5c8020b422fd
Tracked-On: https://jira01.devtools.intel.com/browse/OAM-5604
Signed-off-by: Jean Trivelly <jean.trivelly@intel.com>
Reviewed-on: https://android.intel.com:443/408084
Reviewed-on: https://android.intel.com:443/424900