[Telephony] Removing property:persist.ril-daemon.disable
rild is started in ChargerOS, which is not correct.

As property, persist.ril-daemon.disable is defined as a
ADDITIONAL_DEFAULT_PROPERTIES, it means that it will always be
set in charger mode also. And this property,persist.ril-daemon.disable
is not used to stop the ril daemon. It could be added later, if
needed by DSDS purpose.

Hence removing property:persist.ril-daemon.disable.

Change-Id: I344a2d71a7a27858771454230c8ecc8ed5f29936
Tracked-On: https://jira01.devtools.intel.com/browse/OAM-1683
Signed-off-by: Ramasamy, AlagappanX <alagappanx.ramasamy@intel.com>
Reviewed-on: https://android.intel.com:443/407818