flashfiles/ini: bind the 'check-battery' feature to a group
The 'check-battery' feature might not work if the device does not have
the appropriate firmware release meaning that 'fastboot getvar
battery-voltage' cannot be a mandatory command.  The default value,
must match the minimum battery voltage level.

This patch binds the 'check-battery' feature to a 'check-battery'
group.  The user who do not want the battery level verification is now
able to deactivate it using either the PFT command line either the PFT
UI.

Change-Id: I4a8ba9a66039cbc1f75c7be5aa3ee32864b89475
Tracked-On: https://jira01.devtools.intel.com/browse/OAM-2023
Signed-off-by: Jeremy Compostella <jeremy.compostella@intel.com>
Reviewed-on: https://android.intel.com:443/408561