flashfiles: generic parsing for commands
Now commands fields are directly transformed into json equivalent
without restrictions on the name of the field.
This allows to give generically more datas to phoneflashtool, needed for
the new CSE DnX tool which have to be supported for BXT.
Before, "arg" ini field was converted to "args" json field. Now that the
conversion is generic, we have to rename "arg" in "args" for the ini files.

Tracked-On: https://jira01.devtools.intel.com/browse/OAM-1344
Change-Id: I05d3525a9b03986f740518d73f7fc27d5940fc5e
Signed-off-by: Sylvain Chouleur <sylvain.chouleur@intel.com>
Reviewed-on: https://android.intel.com:443/408889