security: correct shell script execution and policy paths
Launching shell scripts via /system/bin/sh <script> has
the unintentded consequence of having the calling domain
also need to start the interpreter. Another issue is the
clarity of the logs, comm is set to the shell rather than
the script. This fix adjusts that.

Adjust BoardConfig.mk's to add policy locations as well for
various mixins.

Change-Id: I5f96a880b00d31172ed93a4ca908974f17a2c669
Tracked-On: https://jira01.devtools.intel.com/browse/OAM-3211
Signed-off-by: William Roberts <william.c.roberts@intel.com>
Reviewed-on: https://android.intel.com:443/411560