DEXPREOPT: add an option to always enable dexpreopt
It's sometimes very inconvenient for pre-silicon platforms
to use userdebug build due to the fact that dm_verity is
enabled by default, and reboot is required to disable verity.
However using eng build would introduce a huge boot time delay
due to dex2oax conversion when boot for the firsi time.
This patch creates a new option to turn on dexpreopt for all
build variants.

Change-Id: Id89622e962551c2edea5b2f48ffc9e73e237a2bd
Tracked-On: https://jira01.devtools.intel.com/browse/OAM-4166
Signed-off-by: Fei Yang <fei.yang@intel.com>
Reviewed-on: https://android.intel.com/417705
Reviewed-on: https://android.intel.com:443/417705