Autodetect LCD density with a separate executable
In order to achieve a good security design, we should
try to avoid checking device nodes in init process.
Graphics area is a higher risk due to its complexity
and big attack surface. So autodetect LCD density function
is compiled to a separate executable and invoked with
a blocking exec call from the init script.

Change-Id: I5fad88f9897e7da9fcd49c36de375f0d0aed5944
Tracked-On: https://jira01.devtools.intel.com/browse/OAM-1367
Signed-off-by: Alice Liu <alice.liu@intel.com>
Reviewed-on: https://android.intel.com:443/415643