Camera autodetect: Add hald_fuse_atomisp parameter support
This change adds the parameter 'hald_fuse_atomisp' to the camera
autodetect mixin configuration. This parameter is used to enable support
for the atomisp fuse mount point in hald.

With this change, atomisp fuse support is enabled by default for coho
targets and disabled for others with auto-detection enabled but with no
need for atomisp fuse support (bxt and sofia).

Change-Id: I06fe89a873a89723d69667d8d4bef9f7657df79f
Tracked-On: https://jira01.devtools.intel.com/browse/OAM-3836
Signed-off-by: Thierry Escande <thierry.escande@linux.intel.com>
Reviewed-on: https://android.intel.com:443/416185