Replace symlink by copy for modules
Modules are installed in $(KERNEL_MODULES_ROOT) in 2 ways:
First in hierarchical tree to fit kmod requirement that insmod modules
in /lib/modules/`uname -r` ; second in flat structure, to fit with calls
of insmod in init.rc.
Flat structure is symlink to hierarchical tree to not duplicate modules
and save size on disk.
But it happens not to work correctly and insmod is not inserting modules.
Root cause is not explained yet.
This patch is replacing symlink by copy to workaround the issue.

Change-Id: I5c7b5c1108a3b9734331980f78155e722e1a08a7
Tracked-On: https://jira01.devtools.intel.com/browse/OAM-4267
Signed-off-by: Claude Brouat <claude.brouat@intel.com>
Reviewed-on: https://android.intel.com:443/418663