init.rc: Use a separate service directory for Coreu
CoreU service use to create a file 'coreu' in 'data/system' directory
to ensure that only a single instance of CoreU binary is running in
the system. Apparently this was not the right location to create the
file and its recommended to create a separate directory by the service
name through the init.rc and create the file in that directory.
Hence now creating a new directory by name 'coreu' in '/data' and
CoreU service will create the file in this dircetory.
Also the recommendation is to create the directory with owner & group
as 'media', since CoreU eventually runs as a 'media' user. So now
CoreU service is expected to first change its user & group id to
'media' and then access this directory.

Tracked-On: https://jira01.devtools.intel.com/browse/OAM-1368
Change-Id: Ic03d2010fe0b997d2740815f9cf1eebbcdaa22ab
Reviewed-on: https://android.intel.com/418597/
Signed-off-by: Akash Goel <akash.goel@intel.com>
Reviewed-on: https://android.intel.com:443/420814