init.rc: Change the ownership of PCI sysfs file to 'media'
CoreU & hdcpd make use of the Libreg library to access the GFX device
registers. Libreg provides the register IO funtionality by accessing
&mmapping the PCI sysfs file
sys/devices/pci0000:00/0000:00:02.0/resource0'.
This file is accessible to only the Root User, but since now CoreU
& hdcp will open & mmap this file as a 'media' user, need to
accordingly update the ownership of this file to 'media', so as to
completely obviate the need of a DAC_OVERRIDE capability.

Tracked-On: https://jira01.devtools.intel.com/browse/OAM-1368
Change-Id: I70f5e4b6a327141b51f74a43ce7b5a1e107ff8f1
Reviewed-on: https://android.intel.com/418598/
Signed-off-by: Akash Goel <akash.goel@intel.com>
Reviewed-on: https://android.intel.com:443/420822