ModemClient: removal of MMGR implementation
The Java implementation of the modem client library uses a JNI to
communicate with the modem manager. Thanks to this change,
dynamic loading is no more needed and MMGR implementation is
replaced by the JNI implementation.

This patch removes the useless Android properties declaration and
update the list of needed packages.

Change-Id: I51486cec227fc308a4c9dd38b8098dec3014e473
Tracked-On: https://jira01.devtools.intel.com/browse/OAM-5441
Signed-off-by: Cesar DE OLIVEIRA <cesar.de.oliveira@intel.com>
Reviewed-on: https://android.intel.com:443/423725