[CWS_WIDI_UX] BCM43241 Do not scan in background when P2P connected/ing.
When no WiFi network is remembered, wpa_supplicant do not
automatically schedule periodic scan.
In that case, Android framework take the action to trig a scan
with the same interval of time than "suplicant scan interval".
When P2P is connected, this Android framework scan is not triggered
and supplicant scan interval is changed.
Set it to 3600s in order to minimize scan impact on P2P
connection, mainly for Miracast purpose.

Change-Id: Ia906d21fb4f2e41cfc5e8256e5300712b52a983c
Tracked-On:https://jira01.devtools.intel.com/browse/OAM-3717
Signed-off-by: Aymen Ben Alaya <aymenx.ben.alaya@intel.com>
Reviewed-on: https://android.intel.com:443/423899