Copy in image only modules generated by current source tree
During kenrel build, modules are installed in a directory
that contains the commit ID of the head of the source tree.
After successive build of the kernel source tree with modification
committed between each build, many install dir are present.
But only the current one but be copied to the image that
will be flashed on the target

Change-Id: I2f402368f3f81055baacde2bd48177bdd6c2eab1
Tracked-On: https://jira01.devtools.intel.com/browse/OAM-5331
Signed-off-by: Claude Brouat <claude.brouat@intel.com>
Reviewed-on: https://android.intel.com:443/424001