init.rc: Update the ownership of coreu directory & PCI sysfs file to Root
CoreU need to execute the Initialization phase as a Root User,
otherwise the Perfmon functionaility will be affected, as Perfmon
accesses several sysfs & debugfs files which have the write
permission only for the Root User.
It will be cumbersome to update the ownership of all those files
to Media and moreover some of the files are not Graphics specific.
This implies that CoreU would now have to access the /data/coreu
directory and the PCI sysfs file as a Root User only, which will
fail and trigger a dac_override warning, if ownership for the
directory & sysfs file is not updated from Media to Root.
This also implies that the file which CoreU will create inside the
/data/coreu/ directory will also have the Root ownership.

v2: For the /data/coreu directory, update only the group from Media
to Root & keep the owner as Media (William)

Change-Id: If1454088f8410aa7dbcc3b55aaf88ac0dd7f502a
Tracked-On: https://jira01.devtools.intel.com/browse/OAM-7627
Reviewed-on: https://android.intel.com/427542/
Signed-off-by: Akash Goel <akash.goel@intel.com>
Reviewed-on: https://android.intel.com:443/428492