Fixed NVM directory permissions
nvm_fs_partition/nvm is no longer accessible to others as its
permission is set to 0770. In order to still access the nvm, misc is
added to swfg_daemon supplemental groups.
system as the groupname: first parameter of the group options is fine
because the main function the service sets the service gid and uid to
system.

Change-Id: Ic88c54652813ea1e7690d1e878b0449ba8e1c3bc
Signed-off-by: AbdoulayeX Berthe <abdoulayex.berthe@intel.com>
Tracked-On: https://jira01.devtools.intel.com/browse/OAM-4762
Reviewed-on: https://android.intel.com:443/430149