Align the default mv configuration variable name
For Sofia 3GR, the default mv configuration is defined through
MV_CONFIG_DEFAULT_TYPE only. So use the same variable for Sofia LTE.
Moreover, LTE now support multi configuration and as a consequence,
all the configuration files are already added to provdata. There is no
need to add the default one in particular.

Change-Id: I26f64b4e5f35cd730193eb165d12465ebcc1c4e1
Signed-off-by: Gaelle Nassiet <gaellex.nassiet@intel.com>
Tracked-On: https://jira01.devtools.intel.com/browse/OAM-4814
Reviewed-on: https://android.intel.com:443/432829