[CWS_WIDI_UX] BCM43241 Adapter reconnection upon AP just connected on another frequency
Since some WiFi chip has some limitation to offer high video quality while used
in muli-roles and multi-channels, the best solution is to use the same channel as
the access point when running in multi-roles.

This solution is already in place when DUT is first connected to an access point
and then connected to a video adapter. However, in some case we can imagagine that
user first connects to video adapter and then connects to an access point. In
such situation, there is a high probability that adapter and access point
doesn't run on the same channel.

This patch detects when DUT is reaching an access point with a different
frequency than video adapter, a dialog box informs the user that the video link
will be closed and restored for better quality. The user has no other choice
than accept and the patch manages the disconnection, retrieves the adapter we
left and finally performs the reconnection.

Change-Id: I8186eaa4788420fd9a496a67f302baa483a89bf3
Tracked-On: https://jira01.devtools.intel.com/browse/OAM-6992
Signed-off-by: Aymen Ben Alaya <aymenx.ben.alaya@intel.com>
Reviewed-on: https://android.intel.com:443/430210