audio: Set number of warmup cycles in audioflinger
In order to reduce the cold output latency on Baytrail devices,
a property has been set to change the warmup time in the FastMixer.

The property is set to 3 but it can be set from 2 to 10 cycles.
The larger number of cycles, the longer delay is when pushing data
into the fast track.

Change-Id: Id25b55e5ae99092a597970327f1a1059af684fb7
Tracked-On: https://jira01.devtools.intel.com/browse/OAM-6477
Signed-off-by: Alejandro Ochoa <alejandro.ochoa@intel.com>
Reviewed-on: https://android.intel.com:443/435118