autodetect: rework bind-mount and fuse symlinks creation
This modifies the way auto-detection symlinks are defined and created.
link:target pairs are now declared using an absolute path on the android
file system. The links are still created using relative paths but this
relative path is now automatically computed by the link creation rule to
go up to the root fs.

The link:target pair declared as follow:
/system/etc/audio_hw.xml:/system/rt/hal_fuse/audio_hw.xml

will create the link as:
/system/etc/audio_hw.xml -> ../../system/rt/hal_fuse/audio_hw.xml

Change-Id: I0873eae7eed1fb261fe2e72960998d7520444658
Tracked-On: https://jira01.devtools.intel.com/browse/OAM-8923
Signed-off-by: Thierry Escande <thierry.escande@linux.intel.com>
Reviewed-on: https://android.intel.com:443/436958