Add local extrafiles via [extrafiles] keyword in mixins.spec
In order to add different extrafiles for each target,
extand use of [extrafiles] keyword in mixins.spec.
In addtion to [mapping], it allows to select any extra file
and copy it to target dir with a different name.
If a file is in [extrafiles] section of mixins.spec,
mixin-update checks if it is not already in the known list
built with _FILE_LIST and [extrafiles] from files.spec.
It is interesting to choose a specific configuration for a
given target.
If target T1 and T2 need a config file named config.xml based
on config001.xml files and target T3 a config file named config.xml too
based on config002.xml files, then :
Target T1 and T2 should have mixins.spec like::
  [extrafiles]
  config_001.xml: configuration #001
  [mapping]
  config_001.xml: config.xml
Target T3 should have:
  [extrafiles]
  config_002.xml: configuration #002
  [mapping]
  config_002.xml: config.xml
To select a specific file, it is possible to add path to the file to
copy and path to destination file (directory must exist).
To copy only config file in audio/super/config_002.xml in
target_dir/audio, do:
  [extrafiles]
  audio/super/config_002.xml: configuration #002
  [mapping]
  audio/super/config_002.xml: audio/config.xml

Change-Id: I41a76bc642347d8d8469ae0dfb36b6dc61fcdcf0
Tracked-On: https://jira01.devtools.intel.com/browse/OAM-14532
Signed-off-by: Claude Brouat <claude.brouat@intel.com>
Reviewed-on: https://android.intel.com:443/450359