Correct the ro.dalvik.vm.native.bridge value for some devices
The PRODUCT_DEFAULT_PROPERTY_OVERRIDES cannot be overridden for some
devices. Therefore, the ro.dalvik.vm.native.bridge property cannot be
set to libhoudini.so. And houdini is disabled in the image of these
devices.
To solve this issue, the ro.dalvik.vm.native.bridge must be set as the
first entry of PRODUCT_DEFAULT_PROPERTY_OVERRIDES in device makefile.
Then, its default value can be overridden.

Change-Id: I81bfce7bd83e37eafb4adb7c1d7f439268691954
Tracked-On: https://jira01.devtools.intel.com/browse/OAM-12991
Signed-off-by: jgu21 <jinghui.gu@intel.com>
Reviewed-on: https://android.intel.com:443/450191