Add mixins group for captive_portal_detection feature
When connecting to AP, the system checks if captive portal server is defined,
so to check if there is internet access or not.
If not defined, default server “connectivitycheck.gstatic.com” will be used.
This server is blocked in China, so Chinese users always fail the check.
With this issue, there will be always “no internet access” notification
when connecting to an AP (even if the access point has internet access) and
auto connection will be disabled.

To solve this problem, we allow to change captive portal server to "g.cn"
in this patch, Chinese users will be directed to Google's server in Beijing,
and the check will be successful.

This does no harm to those who aren't in China,
because they will be directed to nearest Google server.

This patch adds a mixins groups for captive_portal_detection feature.

Change-Id: I237e94adf2eb2903cf5c057bc313e6a3f094f7d9
Orig-Change-Id: I29411c6cc7db27127d25c61ac4c16af4cf9aac69
Tracked-On: https://jira01.devtools.intel.com/browse/OAM-32132
Signed-off-by: Riadh Khaldi <riadhx.khaldi@intel.com>
Signed-off-by: chuangx <chengcaix.huang@intel.com>
Reviewed-on: https://android.intel.com:443/509249