debug-mpm: export module name to selinux
coreu uses debug-mpm defined types when debug-mpm is enabled,
namely the msr sysfs files. This causes a compilation issue
when coreu (ufo graphics) is enabled but debug-mpm is set to
false.

Correct this by introducing a way to export module names and
ifelse compile around them in the selinux policy.

Change-Id: Iabe07cde49c07147fadc1552d8cb4a797e36bc08
Tracked-On: https://jira01.devtools.intel.com/browse/OAM-13485
Signed-off-by: William Roberts <william.c.roberts@intel.com>
Reviewed-on: https://android.intel.com:443/451503