config-partition: sepolicy refactor
Currently the config partition policy rules are always
being built into the sepolicy even when it is disabled.

Enable the config-partition mixin to use BOARD_SEPOLICY_DIRS
to search for mixin specific policy and export the module
via BOARD_SEPOLICY_M4DEFS, so modules being built can
remove their config-partition rules when this mixin is not
enabled.

Becuase of dependencies in the build of sepolicy, we also
set wlan/bcm43241 and wlan/bcm4356 to include the proper
policy to prevent a build breakage.

Change-Id: Ia310a81b5ec4cf671f656029299ce3f643c65d4f
Tracked-On: https://jira01.devtools.intel.com/browse/OAM-15072
Signed-off-by: William Roberts <william.c.roberts@intel.com>
Reviewed-on: https://android.intel.com:443/455385