Revert "Adjust and copy the audio_effects.conf into target right path."
"r2_cht_cr_m" build is broken because of patch
https://android.intel.com/#/c/452389

The PRODUCT_COPY_FILE copies unconditionally the file in "system/vendor".
This is not good because when the vendor partition is activated
the file should go in "/vendor" instead of "/system/vendor".
"r2_cht_cr_m" target activates the vendor partition,
this is why the build fails here.

It is preferable to define a LOCAL_MODULE here (using BUILD_PREBUILT)
and to add it to a PRODUCT_PACKAGE.

This reverts commit 3a47adcb9f9e4993cd52d94831841ea8851e2862.

Change-Id: I085306bb15762582553b313756425f8fe6b8f7f4
Tracked-On: https://jira01.devtools.intel.com/browse/OAM-15294
Signed-off-by: Di Folco, Neil <neil.di.folco@intel.com>
Reviewed-on: https://android.intel.com:443/455843