security: setup build variables for sepolicy
security_icdg was setting up the sepolicy/security subdir
for its build, however, this conflicts with the naming
convention and things under the security mixin location.
Rename security to icdg and create a new security location
for the existing security mixins. Set up the security mixins
to use sepolicy/security during the build.

Change-Id: I3d0cc249ef334dcaf159e4efe1dbd176ceba6486
Tracked-On: https://jira01.devtools.intel.com/browse/OAM-17459
Signed-off-by: William Roberts <william.c.roberts@intel.com>
Reviewed-on: https://android.intel.com:443/462065