Support conditional compilation of JhsModem.java in JHS
JHS(Java Helper Service) is the service provider to DPTF
in Android. In this patch, a file JhsModem.java is made
to be compiled only if the device is having Intel modem.
This is achieved through adding DPTF in mixin.spec for 'thermal'
with modem related parameter. Based on the value of that
parameter, product.mk defines a variable and it is
checked in Android.mk to add or remove modem related files.

Change-Id: I3b39b0131ddef2b191401cec53cc826bd92bacf0
Tracked-On: https://jira01.devtools.intel.com/browse/OAM-21913
Signed-off-by: Arpana <arpana.bhagwat@intel.com>
Reviewed-on: https://android.intel.com:443/468611