Add Audio Classifier library in PRODUCT_PACKAGE
Audio classifier (also known as Aware in Moorefield
and cherrytrail) is an algorithm responsible for
classifying surrounding audio based on input from
low power digital microphones.
When this algorithm is enabled, it is used as
software audio sensor. The information provided
by algorithm could be used for taking decisions in
android user space.

Change-Id: Id1d25694231beef4d5fa752874191b4421fb0262
Tracked-On: https://jira01.devtools.intel.com/browse/OAM-9905
Signed-off-by: G Kranthi <gudishax.kranthikumar@intel.com>
Reviewed-on: https://android.intel.com:443/475869