Create mixins configuration for app compatibility enhancement feature
This feature is created for some cerner cases of
app compatibility issue. It is disabled by default.
To enable it, please add below lines to mixins.spec.
Currently, it aims to handle below issues:

1. All native libraries are put under assets directory.
Some apps put their libraries under the assets directory
in their APKs, instead of lib directory. It is not a
problem if the ABI of app libs is matched with platform
default ABI. But for apps with other ABIs, this feature
must be enabled to call native bridge for help.

Change-Id: I961b1b959ce002ccdd81bd24186be13e264436b8
Tracked-On: https://jira01.devtools.intel.com/browse/OAM-24201
Signed-off-by: jgu21 <jinghui.gu@intel.com>
Reviewed-on: https://android.intel.com:443/481126