Add the "radio" group to the new AudioServer service
Android N splits the MediaServer service.
Indeed, the AudioFlinger, AudioPolicyManager, FM radio and
SoundTrigger services move to the AudioServer service.

In order to support audio voice use cases, we must add the "radio"
group permission to the AudioServer service.

No need to modify the "audioserver.rc" when building the Kernel only
(i.e. make bootimage).
Indeed, it needs to be patched with radio group, only when building
full image. So we can remove the dependency on the ramdisk,
i.e. removing the module path to TARGET_ROOT_OUT.

Change-Id: Iaebc7401e54f1b9b851e433957618656c1bbda11
Tracked-On: https://jira01.devtools.intel.com/browse/OAM-26838
Signed-off-by: Di Folco, Neil <neil.di.folco@intel.com>
Reviewed-on: https://android.intel.com:443/488272