[VPG]: device/intel/mixins: Start coreu and msync as media user
As per the latest selinux policy enforcement from N-dessert onwards, a
service will not be granted setuid/setgid capabilities except under
extraordinary circumstances.

This arises a need to remove setuid calls from coreu and msync
services. In order to accomplish that both of these services should
start as media user instead of starting as root and then dropping the
privileges later on.

In addition to the above changes we need to add some of the files,
accessed by coreu, to the media user and group. This will ensure the
continued coreu access to these files in its new capabilities.

Change-Id: I54d79af943fe733985172158150d7a16243ea5ad
Tracked-On: https://jira01.devtools.intel.com/browse/OAM-31687
Tracked-On: https://jira01.devtools.intel.com/browse/OAM-31688
Signed-off-by: Praveen Paneri <praveen.paneri@intel.com>
Reviewed-on: https://android.intel.com:443/518506